+++ src/parser/enrichers/class/gunslinger/GutShot.ts
@@ -10,6 +10,7 @@ export default class GutShot extends DDBEnricherData {
     return {
       targetType: "creature",
       targetCount: 1,
+      rangeType: "any",
       activationType: "special",
       activationCondition: "You score a critical hit with a ranged weapon attack against a Large or smaller creature",
       noTemplate: true,
@@ -28,9 +29,11 @@ export default class GutShot extends DDBEnricherData {
         name: "Gut Shot",
         options: {
           durationSeconds: 60,
+          description: "Speed is halved and attack rolls have Disadvantage. Remove this effect when the target replaces an attack with dislodging the projectile.",
         },
         changes: [
           DDBEnricherData.ChangeHelper.movementMultiplierChange("0.5", 20),
+          DDBEnricherData.ChangeHelper.disadvantageAttackChange(),
         ],
       },
     ];